+++ src/diffusers/models/downsamplers.py
@@ -18,9 +18,9 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from ...utils import USE_PEFT_BACKEND
-from ..lora import LoRACompatibleConv
-from .utils import upfirdn2d_native
+from ..utils import USE_PEFT_BACKEND
+from .lora import LoRACompatibleConv
+from .upsamplers import upfirdn2d_native
 
 
 class Downsample1D(nn.Module):
